Add existence check for cache file before reading

+++ src/storage.js
@@ -1,8 +1,11 @@
 const fs = require("fs");
 
 async function getfromcache(id) {
+    if (!fs.existsSync("cache/post/" + id.toString())) {
+        return false;
+    }
+    let content = fs.readFileSync("cache/post/" + id.toString());
     try {
-        let content = fs.readFileSync("cache/post/" + id.toString());
         return JSON.parse(content);
     } catch {
         fs.unlinkSync("cache/post/" + id.toString());
